+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -56,8 +56,8 @@ import {
   colorScheme,
   customBackgroundStyleFn,
 } from '../../src/text-color-style-fn';
-import { getBaseUrl } from '../../src/utils';
 import { linkPreviewUtil } from 'wix-rich-content-common';
+
 import { testWixVideos } from './mock';
 // import { MyCustomIcon, SizeSmallRightIcon, TOOLBARS } from 'wix-rich-content-editor-common';
 import { TOOLBARS, BUTTONS, DISPLAY_MODE } from 'wix-rich-content-editor-common';
@@ -238,7 +238,6 @@ const config = {
   [LINK_PREVIEW_TYPE]: {
     disableEmbed: false,
     fetchMetadata: linkPreviewUtil(authorization),
-    htmlIframeSrc: `${getBaseUrl()}/static/html-plugin-embed.html`,
   },
   [EMOJI_TYPE]: {
     // toolbar: {
@@ -307,7 +306,6 @@ const config = {
     },
   },
   [HTML_TYPE]: {
-    htmlIframeSrc: `${getBaseUrl()}/static/html-plugin-embed.html`,
     minWidth: 35,
     maxWidth: 740,
     width: 350,