+++ examples/main/shared/viewer/ViewerPlugins.jsx
@@ -101,9 +101,7 @@ const config = {
     giphySdkApiKey: process.env.GIPHY_API_KEY,
     sizes: { desktop: 'original', mobile: 'original' }, // original or downsizedSmall are supported
   },
-  [HTML_TYPE]: {
-    htmlIframeSrc: `${getBaseUrl()}/static/html-plugin-embed.html`,
-  },
+  // [HTML_TYPE]: {},
   [LINK_TYPE]: linkPluginSettings,
   [LINK_PREVIEW_TYPE]: {
     disableEmbed: false,