+++ packages/plugin-html/web/src/IframeHtml.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Iframe from './Iframe';
+import htmlIframeSrc from './htmlIframeSrc';
+import styles from '../statics/styles/Iframe.scss';
 
 const isSSR = typeof window === 'undefined';
 
@@ -14,6 +15,16 @@ class IframeHtml extends Component {
     this.setState({ shouldRender: true });
     !isSSR && window.addEventListener('message', this.handleIframeMessage);
   }
+
+  writeToIframe = iframeElement => {
+    this.iframe = iframeElement;
+    const iframeDocument = iframeElement.contentWindow?.document;
+    if (iframeDocument) {
+      iframeDocument.open('text/html', 'replace');
+      iframeDocument.write(htmlIframeSrc);
+      iframeDocument.close();
+    }
+  };
 
   componentWillUnmount() {
     !isSSR && window.removeEventListener('message', this.handleIframeMessage);
@@ -50,7 +61,13 @@ class IframeHtml extends Component {
 
   render() {
     return this.state.shouldRender ? (
-      <Iframe src={this.props.src} iframeRef={this.setIframe} onLoad={this.handleIframeLoad} />
+      <iframe
+        className={styles.iframe}
+        ref={this.writeToIframe}
+        title="remote content"
+        style={{ backgroundColor: 'transparent' }}
+        onLoad={this.handleIframeLoad}
+      />
     ) : null;
   }
 }