+++ packages/plugin-link-preview/web/src/LinkPreviewViewer.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { isEqual } from 'lodash';
 import { mergeStyles, validate, pluginLinkPreviewSchema } from 'wix-rich-content-common';
 import styles from '../statics/styles/link-preview.scss';
+import HtmlComponent from 'wix-rich-content-plugin-html/dist/lib/HtmlComponent';
 
 class LinkPreviewViewer extends Component {
   static propTypes = {
@@ -78,20 +79,18 @@ class LinkPreviewViewer extends Component {
     } = this.styles;
 
     if (!settings.disableEmbed && html) {
-      // const htmlCompProps = {
-      //   componentData: {
-      //     srcType: 'html',
-      //     src: unescape(html),
-      //     ...componentData,
-      //   },
-      //   settings,
-      //   theme,
-      //   isMobile,
-      // };
+      const htmlCompProps = {
+        componentData: {
+          srcType: 'html',
+          src: unescape(html),
+          ...componentData,
+        },
+        settings,
+        theme,
+        isMobile,
+      };
 
-      return <div />;
-
-      // return <HtmlComponent {...htmlCompProps} />;
+      return <HtmlComponent {...htmlCompProps} />;
     } else {
       const { imageRatio } = this.state;
       if (!imageRatio) {