+++ plugins-bundle-analyzer/src/viewer-without-wrapper.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { htmlTypeMapper, HTML_TYPE } from 'wix-rich-content-plugin-html/dist/module.viewer';
+import { htmlTypeMapper } from 'wix-rich-content-plugin-html/dist/module.viewer';
 
 import { RichContentViewer } from 'wix-rich-content-viewer';
 import { imageTypeMapper } from 'wix-rich-content-plugin-image/dist/module.viewer';
@@ -8,10 +8,5 @@ import { linkTypeMapper } from 'wix-rich-content-plugin-link/dist/module.viewer'
 const typeMappers = [imageTypeMapper, linkTypeMapper, htmlTypeMapper];
 
 export default () => {
-  const config = {
-    [HTML_TYPE]: {
-      htmlIframeSrc: `/static/html-plugin-embed.html`,
-    },
-  };
-  return <RichContentViewer typeMappers={typeMappers} config={config} />;
+  return <RichContentViewer typeMappers={typeMappers} />;
 };